perf: 8x32-bit limb field arithmetic (#88)
# 8x32-bit Limb Field Arithmetic

## Performance

NVIDIA GPUs have no native 64-bit integer multiply — a 64x64 multiply is
emulated. The field arithmetic used 4x64-bit saturated limbs, so every
`modular_multiplication` paid for emulated 64x64 products. Switching to
**8x32-bit** limbs makes each partial product a native `32x32 -> 64`
multiply.

Measured on an RTX 5090 (interleaved 64-bit vs 32-bit, same session):

| Metric | 64-bit (main) | this PR |
| --- | --- | --- |
| `bench_modular_multiplication` (dependent chain) | ~24.1 G modmul/s |
**~27.3 G modmul/s (+13%)** |
| `bench_ckdpub_throughput` | ~78.2 M keys/s | **~85.1 M keys/s
(+8.8%)** |

End-to-end `--gpu-only` on the same GPU went from ~71 M to ~78 M addr/s.

Note: squaring is currently `uint256_multiplication(a, a)` (like
hashcat/BitCrack, which have no dedicated square). A dedicated 8x32
square is a separate follow-up so its incremental gain can be measured
on its own.

## Main Changes

Representation is **big-limb-first** (`limbs[0]` most significant), same
convention as before, now `uint[8]` / `uint[16]`.

### Field layer
- `Uint256` -> `uint[8]`, `Uint512` -> `uint[16]`; `Uint320` retired.
- Byte conversion (`UINT256_FROM_BYTES`/`uint256_to_bytes`, and 512)
rewritten for 8/16 limbs.
- `uint256_addition`/`subtraction`/`shift` rewritten as 8-limb
carry/borrow chains.
- **Multiply:** Comba (column-scanning) 8x8 with a `ulong` accumulator +
`uint` carry counter (portable OpenCL, no inline asm — Intel Arc still
works). Each column sums <= 8 products < 2^67; validated against the bit
bounds.
- **Reduction:** fold `H*c = (H << 32) + H*977` with `c = 2^256 mod p =
2^32 + 977` (the `<< 32` is a free limb shift with 32-bit limbs), a
second fold of the ~33-bit remainder, then one conditional subtraction
of p. Structure matches hashcat and BitCrack (verified against both).
Retired the old `uint256_ulong_multiplication`.
- Constants: `SECP256K1_P_0..7`; removed dead `SECP256K1_G*`/`P_MINUS_*`
macros and the old folding constant.

### EC layer
- `g_times_scalar` digit loop: 8 limbs x shifts 24..0.
- `jacobian_point_affine_point_addition`: infinity checks and P
comparison over 8 limbs; the z=1 constant is `{0,...,0,1}` (8 limbs).
- `ckdpub` / `compress_point`: y parity reads `limbs[7]` (LSW).
- `modular_addition`/`subtraction`/`double`: "subtract p" comparison
chains extended to 8 limbs.

### Rust (GPU buffer producers, in lockstep with the struct layout)
- `gpu_cache.rs` `Uint256 { limbs: [u32; 8] }`, `g_tables.rs` and
`cache_preloader.rs` build 8x u32 from big-endian bytes.

## Testing

- All existing byte-interface vectors still pass
(representation-independent).
- Added 32-bit-boundary vectors for multiply, reduction (`2^511 mod p`),
and the modular ops that stress carries across the new 4-byte limb
boundaries.
- Retired the tests for the removed ops (`uint256_ulong_multiplication`,
`uint320_uint256_addition`).
- Full suite on RTX 5090: **388 passed, 0 failed** (serialized) —
includes the BIP32 `ckdpub` vectors and `batch_address_search` vs CPU,
which guards the Rust<->GPU buffer layout.

## Next Steps

- Dedicated 8x32 squaring (measured on its own).
- Vendor-gated NVIDIA inline-PTX carry-chain (`mad.lo.cc`/`madc.hi.cc`)
for the multiply/reduction, with the portable path as fallback.

+++ src/opencl/kernels/g_times_scalar_compute_kernel.cl
@@ -11,7 +11,7 @@ __kernel void g_times_scalar_compute_kernel(
     __global const Point *g_times_tables)
 {
     int thread_id = get_global_id(0);
-    ulong offset = *iteration_offset;
+    uint offset = (uint)(*iteration_offset);
 
     // Copy data from global to private memory and convert
     uchar scalar_private[32];
